TST: split tests and cover some more cases
